+++ CHANGELOG.md
@@ -2,6 +2,33 @@
 
 The versions of the sovity EDC CE are aligned with the sovity EDC EE.
 
+## [v16.7.0] - 2026-08-19
+
+### Overview
+
+Allow custom vault key for Azure Blob transfers.
+
+#### Minor Changes
+
+- Add new required argument for custom vault key name for Azure Blob transfers in UI and UI API
+
+#### Patch Changes
+
+- Remove `*.complete´ marker file creation for Azure Blob transfers
+
+### Deployment Migration Notes
+
+_No special deployment migration steps required_
+
+#### Compatible Versions
+
+- EDC CE Backend: `ghcr.io/sovity/edc-ce:16.7.0`
+- EDC CE Frontend: `ghcr.io/sovity/edc-ce-ui:16.7.0`
+- PostgreSQL: `17`
+- Eclipse EDC Fork: [v0.11.1.3](https://github.com/sovity/core-edc/releases/tag/v0.11.1.3)
+- Tractus-X: `0.9.0`
+
+
 ## [v16.6.2] - 2026-07-07
 
 ### Overview